Move ps/2
