+++ kernel/devices/libdevices.cpp
@@ -1,0 +1,2 @@
+
+// an empty source to make CMake happy
