+++ kernel/devices/ps2/include/ps2/common.hpp
